requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==24.3.25
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
keras==3.1.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.7
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.2.0
protobuf==4.25.3
Pygments==2.17.2
pytesseract==0.3.10
requests==2.31.0
rich==13.7.1
segment-anything==1.0
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
torch==2.2.1
torchvision==0.17.1
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
